REF: move isIOS26OrHigher and settings header options out of platform layer
