+++ blue_modules/environment.ts
@@ -1,7 +1,10 @@
+import { Platform } from 'react-native';
 import { getDeviceType, isTablet as checkIsTablet } from 'react-native-device-info';
 
 const isTablet: boolean = checkIsTablet();
 const isDesktop: boolean = getDeviceType() === 'Desktop';
 const isHandset: boolean = getDeviceType() === 'Handset';
 
-export { isDesktop, isHandset, isTablet };
+const isIOS26OrHigher: boolean = Platform.OS === 'ios' && parseInt(String(Platform.Version), 10) >= 26;
+
+export { isDesktop, isHandset, isIOS26OrHigher, isTablet };
